Add Cucumber test suite
