+++ Cargo.toml
@@ -12,3 +12,7 @@ hex = "0.4"
 chrono = "0.4"
 rand = "0.8"
 data-encoding = "2.*"
+
+[dev-dependencies]
+cucumber = "0.21"
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
